+++ Programming/pic_uC/controller.X/thermistor.c
@@ -8,12 +8,13 @@
 
 #include "includes.h"
 
-double read_V(void) {
+float read_V(void) {
   return ((read_adc()*VCC)/1024);
+  /* return (read_adc()*READ_V_CONSTANT) */
 }
 
-double read_V_averaged(void) {
-  double sum = 0;
+float read_V_averaged(void) {
+  float sum = 0;
   uint8_t i;
   for(i=0; i<NUM_SAMPLES; i++) {
     sum += read_V();
@@ -23,14 +24,14 @@ double read_V_averaged(void) {
 
 int read_R_averaged(void) {
 
-  double Vi = read_V_averaged();
+  float Vi = read_V_averaged();
   return (int)((R_10K*Vi) / (VCC-Vi));
 
 }
 
-double read_T() {
+float read_T() {
 
-  double T_inverse = ROOM_TEMPERATURE_INV + B_FACTOR_INV*log(read_R_averaged()/R_NOMINAL);
+  float T_inverse = ROOM_TEMPERATURE_INV + B_FACTOR_INV*log(read_R_averaged()/R_NOMINAL);
   return ((1/T_inverse) - 273.15);
 
 }
